+++ src/coding/mod.rs
@@ -1,8 +1,8 @@
 use std::collections::VecDeque;
 
 pub mod boolean;
-pub mod ibyte;
-pub mod ubyte;
+pub mod signed_byte;
+pub mod unsigned_byte;
 pub mod difficulty;
 pub mod dimension;
 pub mod gamemode;
